+++ src/interfaces/IEtherFiOracle.sol
@@ -13,7 +13,7 @@ interface IEtherFiOracle {
         uint256[] liquidityPoolValidatorsToExit;
         uint256[] exitedValidators;
         uint256[] slashedValidators;
-        uint256[] withdrawalRequestsToInvalidate;
+        uint32[] withdrawalRequestsToInvalidate;
         uint32 lastFinalizedWithdrawalRequestId;
         uint32 eEthTargetAllocationWeight;
         uint32 etherFanTargetAllocationWeight;
