+++ src/interfaces/ILiquidityPool.sol
@@ -50,9 +50,9 @@ interface ILiquidityPool {
     function deposit(address _user, bytes32[] calldata _merkleProof) external payable returns (uint256);
     function deposit(address _user, address _recipient, bytes32[] calldata _merkleProof) external payable returns (uint256);
     function withdraw(address _recipient, uint256 _amount) external returns (uint256);
-    function requestWithdraw(address recipient, uint256 amount) external returns (uint256);
-    function requestWithdrawWithPermit(address _owner, uint256 _amount, PermitInput calldata _permit) external returns (uint256);
-    function requestMembershipNFTWithdraw(address recipient, uint256 amount) external returns (uint256);
+    function requestWithdraw(address recipient, uint256 amount) external returns (uint32);
+    function requestWithdrawWithPermit(address _owner, uint256 _amount, PermitInput calldata _permit) external returns (uint32);
+    function requestMembershipNFTWithdraw(address recipient, uint256 amount) external returns (uint32);
 
     function batchDepositAsBnftHolder(uint256[] calldata _candidateBidIds, uint256 _index, uint256 _numberOfValidators) external payable returns (uint256[] memory);
     function batchRegisterAsBnftHolder(bytes32 _depositRoot, uint256[] calldata _validatorIds, IStakingManager.DepositData[] calldata _registerValidatorDepositData, bytes32[] calldata _depositDataRootApproval, bytes[] calldata _signaturesForApprovalDeposit) external;
